move time-converter methods from Utils to TimeHelper
