+++ src/helpers/TimeHelper/TimeHelper.js
@@ -1,5 +1,26 @@
 import moment from 'moment';
 
+function hoursToHoursAndMinutes (hours) {
+    const totalMinutes = hours * 60;
+
+    const convertedHours = Math.floor(totalMinutes / 60);
+    let convertedMinutes = Math.floor(totalMinutes % 60);
+
+    if (convertedMinutes.toString().length == 1) {
+        convertedMinutes = '0' + convertedMinutes;
+    }
+
+    return `${convertedHours}:${convertedMinutes}`;
+};
+
+function hoursAndMinutesToHours (time) {
+    time = time.toString();
+    const timeComponents = time.split(':');
+    const dec = parseInt((timeComponents[1]/6)*10, 10);
+
+    return parseFloat(parseInt(timeComponents[0], 10) + '.' + (dec<10?'0':'') + dec);
+};
+
 function iso8601ToDDMMYYY (date) {
     const isDDMMYYYY = moment(date, 'DD.MM.YYYY').format('DD.MM.YYYY') === date;
 
@@ -21,6 +42,8 @@ function ddMMYYYYToISO8601 (date) {
 };
 
 export {
+    hoursToHoursAndMinutes,
+    hoursAndMinutesToHours,
     iso8601ToDDMMYYY,
     ddMMYYYYToISO8601
 };
